Move test files to lib dir
