+++ lib/models/config.test.ts
@@ -1,4 +1,4 @@
-import { crowi, Fixture } from '../setup'
+import { crowi, Fixture } from 'server/test/setup'
 
 describe('Config model test', () => {
   let Config
