+++ lib/test/crowi-environment.js
@@ -1,8 +1,8 @@
 const NodeEnvironment = require('jest-environment-node')
 const MongodbMemoryServer = require('mongodb-memory-server').default
 const path = require('path')
-const ROOT_DIR = path.join(__dirname, '../')
-const MODEL_DIR = path.join(__dirname, '../lib/models')
+const ROOT_DIR = path.join(__dirname, '../..')
+const MODEL_DIR = path.join(__dirname, '../models')
 
 class CrowiEnvironment extends NodeEnvironment {
   constructor(config) {
